Test new (to me) style of git commit messages
I recently read an article on how to style commit messages
to maximizxe their utility, but I've never written a commit
message with a subject and body before, so I'm trying that
out now in this meaningless commit on this school project
rather than figure it out in a scenare where the repo is
actually important to a business. I mean, I've already del-
eted and remade this one... three times? Something like that