Implement Scraper and fix many unti test bugs
Not all tests pass. In the case of #scrape_profile, this is because of
bugs in the test and not the code. In the case of #scrape-league_list
it is certainly both.

Currently Scraper builds the entirety of every object in one giant
nested hash. That will almost certainly require splitting into separate
parts once I start implementing the objects.